INTPYTHON-649 Rename the project to pymongo-voyageai-multimodal
