+++ pymongo_voyageai_multimodal/_version.py
