+++ pymongo_voyageai_multimodal/document.py
