+++ pymongo_voyageai_multimodal/storage.py
